Fix mojo logger usage
The maven plugin documentation [1] and the javadoc of the `getLog()` method
imply that it is not recommend to cache the logger on instance creation.

This way an default console logger is used instead of the correctly configured
logger. Which leads to issues when the log level is changed, for example
by invoking the plugin with `mvn -q` to prevent output. The plugin will
still create output.

This change replaces all usages of the cached logger by calling `getLog()`
as recommend by the javadoc.
>  Therefore, simply call this method directly whenever you need the logger, it is fast enough and needs no caching.

[1] https://maven.apache.org/plugin-developers/common-bugs.html#Retrieving_the_Mojo_Logger

fixes #30