Suggest google-java-format integration for IDE's
Instead of linking to Google Style-compliant configurations for those IDEs.

The behavior of, e.g.: Eclipse configured with `GoogleStyle` and `google-java-format` are subtly different in ways that can cause confusion (see google/google-java-format#504). Using the normal integration of the plugin will ensure consistent formatting in the IDE and the format plugin here.